Bitcoin generation renamed package
