+++ btc-address-generation/build.gradle
@@ -13,26 +13,26 @@ apply plugin: 'application'
 apply plugin: "kotlin-spring" // See https://kotlinlang.org/docs/reference/compiler-plugins.html#kotlin-spring-compiler-plugin
 apply plugin: 'org.springframework.boot'
 apply plugin: 'com.github.johnrengelman.shadow'
-mainClassName = "generation.btc.BtcAddressGenerationMain"
+mainClassName = "com.d3.btc.generation.BtcAddressGenerationMain"
 
 dependencies {
     compile project(":btc")
 }
 
 task runBtcAddressGeneration(type: JavaExec) {
-    main = 'generation.btc.BtcAddressGenerationMain'
+    main = 'com.d3.btc.generation.BtcAddressGenerationMain'
     classpath = sourceSets.main.runtimeClasspath
     setWorkingDir("$rootDir/")
 }
 
 task generateBtcChangeAddress(type: JavaExec) {
-    main = 'generation.btc.trigger.BtcChangeAddressGeneration'
+    main = 'com.d3.btc.generation.trigger.BtcChangeAddressGeneration'
     classpath = sourceSets.main.runtimeClasspath
     setWorkingDir("$rootDir/")
 }
 
 task generateBtcFreeAddress(type: JavaExec) {
-    main = 'generation.btc.trigger.BtcFreeAddressGeneration'
+    main = 'com.d3.btc.generation.trigger.BtcFreeAddressGeneration'
     classpath = sourceSets.main.runtimeClasspath
     setWorkingDir("$rootDir/")
 }
