+++ btc-address-generation/src/main/kotlin/com/d3/btc/generation/config/BtcAddressGenerationAppConfiguration.kt
@@ -1,4 +1,4 @@
-package generation.btc.config
+package com.d3.btc.generation.config
 
 import com.d3.btc.wallet.WalletFile
 import config.loadConfigs
