+++ btc-address-generation/src/main/kotlin/com/d3/btc/generation/main.kt
@@ -1,13 +1,13 @@
 @file:JvmName("BtcAddressGenerationMain")
 
-package generation.btc
+package com.d3.btc.generation
 
 import com.github.kittinunf.result.Result
 import com.github.kittinunf.result.failure
 import com.github.kittinunf.result.flatMap
 import config.getProfile
-import generation.btc.config.btcAddressGenerationConfig
-import generation.btc.init.BtcAddressGenerationInitialization
+import com.d3.btc.generation.config.btcAddressGenerationConfig
+import com.d3.btc.generation.init.BtcAddressGenerationInitialization
 import mu.KLogging
 import org.springframework.boot.SpringApplication
 import org.springframework.boot.autoconfigure.SpringBootApplication
@@ -16,11 +16,11 @@ import org.springframework.context.annotation.ComponentScan
 @SpringBootApplication
 @ComponentScan(
     basePackages = [
-        "generation.btc",
+        "com.d3.btc.generation",
         "com.d3.btc.healthcheck",
         "com.d3.btc.provider.generation",
         "com.d3.btc.provider.network",
-        "generation.btc.trigger"]
+        "com.d3.btc.generation.trigger"]
 )
 class BtcAddressGenerationApplication
