+++ btc-address-generation/src/main/kotlin/com/d3/btc/generation/trigger/ChangeAddressGeneration.kt
@@ -1,6 +1,6 @@
-@file:JvmName("BtcFreeAddressGeneration")
+@file:JvmName("BtcChangeAddressGeneration")
 
-package generation.btc.trigger
+package com.d3.btc.generation.trigger
 
 import com.d3.btc.model.BtcAddressType
 import com.github.kittinunf.result.Result
@@ -9,22 +9,23 @@ import mu.KLogging
 import org.springframework.context.annotation.AnnotationConfigApplicationContext
 import org.springframework.context.annotation.ComponentScan
 
-@ComponentScan(basePackages = ["generation.btc.trigger"])
-class BtcFreeAddressGenerationApplication
+@ComponentScan(basePackages = ["com.d3.btc.generation.trigger"])
+class BtcChangeAddressGenerationApplication
 
 private val logger = KLogging().logger
 /*
-   Free address generation entry point
+   Change address generation entry point
  */
 fun main(args: Array<String>) {
-    val addressType = BtcAddressType.FREE
+    val addressType = BtcAddressType.CHANGE
     Result.of {
-        AnnotationConfigApplicationContext(BtcFreeAddressGenerationApplication::class.java)
+        AnnotationConfigApplicationContext(BtcChangeAddressGenerationApplication::class.java)
     }.flatMap { context ->
-        context.getBean(AddressGenerationTrigger::class.java).startAddressGeneration(
-            addressType = addressType,
-            nodeId = btcAddressGenerationTriggerConfig.nodeId
-        )
+        context.getBean(AddressGenerationTrigger::class.java)
+            .startAddressGeneration(
+                addressType = addressType,
+                nodeId = btcAddressGenerationTriggerConfig.nodeId
+            )
     }.fold(
         {
             logger.info { "${addressType.title} address generation process was started" }
